extract Transcoder into it's own file, git rename of MessageParser source file
